Split parsers into several files
